firmware/requirements.txt
PIL
serial
readerwriterlock
distro
gi
bitstring==3.1.5
future==0.16.0
gpio==0.2.0
Jinja2==2.10
lazy-import==0.2.2
numpy==1.14.5
paho-mqtt==1.4.0
pyalsaaudio==0.8.4
pydbus==0.6.0
pygame==1.9.4
pygobject==3.28.3
pyserial==3.4
six==1.11.0
systemd-python==234
toml==0.10.0
websocket-server==0.4
mock>=2.0.0
